[Bugfix] Case where height of svg was off by size of one figure
Use Math.ceil(...) instead of (...) + 1 to calculate height

Example case this catches:
- num_figs = 1, FIG_DIM = 250
- height_old = (((1 * 250) / 250) + 1) * 250 = (2 * 250) = 500;
- height_new = Math.ceil((1 * 250) / 250)) * 250 = ( 1 * 250) = 250;

Having accurate w and h is important when we are using saveSvgAsPng.js
to download the svg as a png. I am adding this in an upcoming commit.

Test case:
test_data2.csv
Set NORMALIZE_DATA to true and NORMALIZE_BY to 'person'