[Download] Set filename to start_to_end.png if start & end defined
If start or end is not defined, then save it as 'radar.png'

This was useful for an earlier data pull, where I was asked to
generate and save weekly images for Kasich.

Moved add_tweaks_html to within main render_radars function so that it
has access to the start and end variables.