Clean blatant data. Remove newline char. Download button works now.
Extra ^M character deleted at end of new line. Googling shows it
occurs on Windows? Maybe because it was copied from an Excel file?

Not sure why this was preventing the download though.

Changing & to / was just to make the charts read a little better.
Topics are split by spaces into new lines, so using a / results in
less tall text that encroaches less on the plot. Not that this matters lol.