[Tweaks] Update blobs as slider moves, as opposed to on mouseup
Before, pointiness, fullness, and color opacity were changed
only on mouse up. Making iterative updates was more cumbersome.
Now, blobs dynamically update as the slider moves.

NOTE: Blobs need to be redrawn when tension or intermeidate points
change because the control points and cardinal line change. HOWEVER,
blobs do NOT need to be redrawn when opacity changes. Obviously just a
CSS change. Adding this (just update the CSS) as a TODO.

Note, input.local.on('change', ...) is still required to get
color to update. In upcoming commit, colorpicker will be dynamic as well.