[API] render_radars takes assoc-array with display/data params
Parameters controlling what parts of the radar charts to display, as well
as whether or not to normalize the data, are passed to render_radars. The
defaults are defined in the read_params function. The default is used if
the key is not present in the passed associative array.

This makes radar.js easier to drop into another project.