[API] render_radars takes assoc-array for default tweak parameters
- If key is not in assoc array, use default values
- Default values are in initialize_tweak_params and initialize_colorpicker
- Implementation: add_tweaks_html is called, which 1) adds the html elements,
2) initializes them using the assoc-array, 3) initializes the spectrum
colorpicker, 4) adds the download button. Then, update_tweaks() is called
to read the input/slider values and set the variables. NOTE, this is kind of
kludgy dependence.

Goal of this commit is to make radar.js more plug and play. Motivation was
using this in the dashboard. When updating data using checkboxes, the
tweak parameter settings were not saved, as the plots were wiped and redrawn
with the defaults defined in radar.js Now, however, we can store the state,
and then pass it to radar.js

Test plan:
- Pass in different combinations of defined and undefined tweak parameters,
  check that defaults/params are properly loaded. Check other functionality